fix: Mocks and testhelpers folders moved to the root
